requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
websockets==12.0
httpx==0.27.0
pydantic==2.7.1

# LangChain & LangGraph
langchain==0.2.1
langchain-openai==0.1.8
langchain-core==0.2.5
langgraph==0.1.5

# STT
assemblyai==0.26.0

# TTS
cartesia==1.0.0

# Utilities
anyio==4.4.0
